# api/rest/error.go
// Package api holds HTTP transport DTOs and standard JSON response envelopes
// for the web adapters.
package rest

// Error is the standard JSON error envelope returned by all endpoints.
// For validation failures, Details contains per-field messages keyed by the
// JSON field name.
type Error struct {
	Error   string            `json:"error"             example:"validation failed"`
	Details map[string]string `json:"details,omitempty"`
}